Move handlers into indiviudal files by resource type
