+++ server/api/index.js
@@ -1,0 +1,13 @@
+const express = require('express')
+const search = require('./search')
+const candidate = require('./candidate')
+const contributor = require('./contributor')
+
+const api = express.Router()
+api.use(express.json())
+
+api.use(search)
+api.use(candidate)
+api.use(contributor)
+
+module.exports = api
