+++ server/app.js
@@ -27,7 +27,7 @@ if (process.env.NODE_ENV === 'production') {
   app.use(enforceSSL)
 }
 
-app.use(bodyParser.json())
+app.use(express.json())
 app.use('/api', api)
 app.get('/status', (req, res) => res.send({ status: 'online' }))
